fix: Move to correct package
This breaks API but since its a fix thats fine
