+++ src/main/kotlin/app/revanced/patches/messenger/inbox/HideInboxSubtabsPatch.kt
@@ -1,17 +1,17 @@
-package app.revanced.patches.messenger.inbox.subtabs.patch
+package app.revanced.patches.messenger.inbox
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.replaceInstruction
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
-import app.revanced.patches.messenger.inbox.subtabs.fingerprints.CreateInboxSubTabsFingerprint
+import app.revanced.patches.messenger.inbox.fingerprints.CreateInboxSubTabsFingerprint
 import app.revanced.util.exception
 
 @Patch(
     name = "Hide inbox subtabs",
     description = "Hides Home and Channels tabs between active now tray and chats.",
-    compatiblePackages = [CompatiblePackage("com.facebook.orca")]
+    compatiblePackages = [CompatiblePackage("com.facebook.orca")],
 )
 @Suppress("unused")
 object HideInboxSubtabsPatch : BytecodePatch(
