+++ src/main/kotlin/app/revanced/patches/messenger/inbox/fingerprints/LoadInboxAdsFingerprint.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.messenger.inbox.ads.fingerprints
+package app.revanced.patches.messenger.inbox.fingerprints
 
 import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.MethodFingerprint
@@ -8,11 +8,10 @@ internal object LoadInboxAdsFingerprint : MethodFingerprint(
     returnType = "V",
     strings = listOf(
         "ads_load_begin",
-        "inbox_ads_fetch_start"
+        "inbox_ads_fetch_start",
     ),
     accessFlags = AccessFlags.PUBLIC or AccessFlags.STATIC,
     customFingerprint = { methodDef, _ ->
         methodDef.definingClass == "Lcom/facebook/messaging/business/inboxads/plugins/inboxads/itemsupplier/InboxAdsItemSupplierImplementation;"
-    }
+    },
 )
-
