+++ src/main/kotlin/app/revanced/patches/messenger/inputfield/DisableSwitchingEmojiToStickerPatch.kt
@@ -1,6 +1,5 @@
-package app.revanced.patches.messenger.inputfield.patch
+package app.revanced.patches.messenger.inputfield
 
-import app.revanced.util.exception
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.extensions.InstructionExtensions.replaceInstruction
@@ -8,16 +7,17 @@ import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import app.revanced.patches.messenger.inputfield.fingerprints.SwitchMessangeInputEmojiButtonFingerprint
+import app.revanced.util.exception
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 
 @Patch(
     name = "Disable switching emoji to sticker",
     description = "Disables switching from emoji to sticker search mode in message input field.",
-    compatiblePackages = [CompatiblePackage("com.facebook.orca")]
+    compatiblePackages = [CompatiblePackage("com.facebook.orca")],
 )
 @Suppress("unused")
 object DisableSwitchingEmojiToStickerPatch : BytecodePatch(
-    setOf(SwitchMessangeInputEmojiButtonFingerprint)
+    setOf(SwitchMessangeInputEmojiButtonFingerprint),
 ) {
     override fun execute(context: BytecodeContext) {
         SwitchMessangeInputEmojiButtonFingerprint.result?.let {
@@ -28,7 +28,7 @@ object DisableSwitchingEmojiToStickerPatch : BytecodePatch(
 
                 replaceInstruction(
                     setStringIndex,
-                    "const-string v$targetRegister, \"expression\""
+                    "const-string v$targetRegister, \"expression\"",
                 )
             }
         } ?: throw SwitchMessangeInputEmojiButtonFingerprint.exception
